Move more command logic to core.
